% Generated by roxygen2 (4.0.1): do not edit by hand
\name{df.to.gr}
\alias{df.to.gr}
\title{Convert a data.frame to a GRanges object}
\usage{
df.to.gr(x, which.chr = "chr", which.start = "start", which.end = "end",
  which.width = "width", which.name = "name", chrlength = NULL,
  start0 = TRUE)
}
\arguments{
\item{x}{data.frame or matrix}

\item{which.chr}{character, the column name of `chr`}

\item{which.start}{character, the column name of `start`}

\item{which.end}{character, the column name of `end`}

\item{which.width}{character, the column name of `width`}

\item{which.name}{character, the column name of `name`}

\item{chrlength}{numeric, the lengths of the chromosomes}

\item{start0}{logical, wheter the `start` is 0-based.}
}
\value{
a GRanges object
}
\description{
Convert a data.frame to a GRanges object
}
\examples{
\dontrun{
x <- data.frame(
  chr=c('chr1','chr2'),start=0,end=100,
  name=paste0('ID',1:2),score=1:2
  )
df.to.gr(x)
df.to.gr(x,chrlength=c(chr1=1000,chr2=1200))
}
}
\author{
Xiaobei Zhao
}

