% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngscopy.R
\name{help_segmtype}
\alias{help_segmtype}
\title{Get help for segmentation functions}
\usage{
help_segmtype(segtype)
}
\arguments{
\item{segtype}{character, the type of segmentation.
See \code{parse_segmtype()}}
}
\description{
Get help for segmentation functions given the type of
segmentation.
}
\examples{
\dontrun{
help_segmtype(parse_segmtype()[1])
}
}
\author{
Xiaobei Zhao
}

