% Generated by roxygen2 (4.0.1): do not edit by hand
\name{help_segtype}
\alias{help_segtype}
\title{Get help for segmentation functions}
\usage{
help_segtype(segtype)
}
\arguments{
\item{segtype}{character, the type of segmentation.
See \code{parse_segtype()}}
}
\description{
Get help for segmentation functions given the type of
segmentation.
}
\examples{
\dontrun{
help_segtype(parse_segtype()[1])
}
}
\author{
Xiaobei Zhao
}

