% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngscopy-test.R
\name{ngscopy_cmdline_example}
\alias{ngscopy_cmdline_example}
\title{Run an example of NGScopy at UNIX-like command line}
\usage{
ngscopy_cmdline_example(ifRun = FALSE, pcThreads = 1)
}
\arguments{
\item{ifRun}{logical, whether to run}

\item{pcThreads}{numeric,
the number of processors performing the parallel computing.}
}
\value{
character, the command line.
}
\description{
Run an example of NGScopy at UNIX-like command line
}
\examples{
## To run at R prompt
ngscopy_cmdline_example()
\dontrun{
ngscopy_cmdline_example(ifRun=TRUE) # Note: this would take a while. And
                                    # R will create a folder for output
                                    # at current working directory
}

## ## To run at Unix-like command line (not at R prompt)
## Rscript -e "require(methods);NGScopy::ngscopy_cmdline_example(TRUE)"
}
\author{
Xiaobei Zhao
}

