% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ngscopy_unittest}
\alias{ngscopy_unittest}
\title{A wrapper to run unit testing of NGScopy}
\usage{
ngscopy_unittest(...)
}
\arguments{
\item{...}{additional arguments to \code{UnitTest$new}}
}
\description{
A wrapper to run unit testing of NGScopy
}
\examples{
## To run at R prompt
\dontrun{
ngscopy_unittest() # Note: R will create a folder for output
                   # at current working directory
}

## ## To run at Unix-like command line (not at R prompt)
## Rscript -e "require(methods);NGScopy::ngscopy_unittest()"
}
\author{
Xiaobei Zhao
}

