% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parse_segmtype}
\alias{parse_segmtype}
\title{Parse the type of segmentation}
\usage{
parse_segmtype(segtype)
}
\arguments{
\item{segtype}{character, the type of segmentation.
Return all available types if missing.}
}
\value{
the type of segmentation
}
\description{
Parse the type of segmentation or return all available types
with missing input.
}
\examples{
parse_segmtype()
}
\author{
Xiaobei Zhao
}

