% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngscopy.R
\name{read_regions}
\alias{read_regions}
\title{Read regions from a data.frame, a file or a connection.}
\usage{
read_regions(x)
}
\arguments{
\item{x}{data.frame or character, the regions in study.
A data.frame with three columns (chr/start/end), a file path or a
connection with three columns (chr/start/end) without column names.
Return an empty data.frame if "", "NULL" or NULL.}
}
\value{
data.frame of three columns (chr/start/end)
}
\description{
Read regions from a data.frame, a file or a connection.
}
\examples{
read_regions("
chr1 0 249250621
chr2 0 243199373
chr3 0 198022430
chr4 0 191154276
chr5 0 180915260
chr6 0 171115067
")
}
\author{
Xiaobei Zhao
}

