% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NGScopy-package.R
\docType{package}
\name{NGScopy-package}
\alias{NGScopy-package}
\title{NGScopy: Detection of copy number variations in next generation sequencing}
\description{
NGScopy provides a quantitative caller for detecting copy number variations in next generation sequencing (NGS), including whole genome sequencing (WGS), whole exome sequencing (WES) and targeted panel sequencing (TPS). The caller can be parallelized by chromosomes to use multiple processors/cores on one computer.
}
\author{
Xiaobei Zhao
}
\references{
Zhao et al (2014), Targeted Sequencing in Non-Small Cell Lung Cancer (NSCLC) Using the University of North Carolina (UNC) Sequencing Assay Captures Most Previously Described Genetic Aberrations in NSCLC, In preparation
}
\keyword{package}

